Rename johnson-solids-challenge-models.js to johnson-solids-models.js
